backend/requirements.txt
# Multi-Agent Framework
crewai>=0.80.0
crewai-tools>=0.12.0

# OpenAI (required by CrewAI)
openai>=1.0.0

# Vector Database (for RAG memory)
chromadb>=0.4.0
sentence-transformers>=2.2.0  # For embeddings

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP Client (for Supabase REST API)
requests>=2.31.0

# PostgreSQL Database Driver
psycopg2-binary>=2.9.9

# Utilities
python-dateutil>=2.8.2
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0  # For API testing